Time: 32 ms (50.58%), Space: 106.7 MB (83.06%) - LeetHub

+++ 2126-destroying-asteroids/2126-destroying-asteroids.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool asteroidsDestroyed(int mass, vector<int>& asteroids) {
         int n = asteroids.size();
-        long long m = (long long)(mass);
+        long long m = mass;
         sort(asteroids.begin(), asteroids.end());
         for(int i = 0; i < n; i++){
             if(m <= 0) return false;
